requirements.txt
langchain-google-genai
python-dotenv
pytest